+++ app/src/main/java/com/example/marketplace_mvp/screens/TitleScreen.kt
@@ -13,21 +13,7 @@ class TitleScreen : Fragment(R.layout.title_screen) {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-
-        // Обработчик клика по всей области (как кнопка)
-        view.setOnClickListener {
-            navigateToDefaultScreen()
-        }
-
-        // Меняем курсор на кликабельный
-        view.isClickable = true
-        view.foreground = ContextCompat.getDrawable(requireContext(), android.R.drawable.btn_default)
     }
 
-    private fun navigateToDefaultScreen() {
-        val appSettings = AppSettings(requireContext())
-        appSettings.setFirstLaunchCompleted()
-        findNavController().navigate(R.id.action_titleScreen_to_toolBarFragment)
-    }
 }
