+++ app/src/main/res/navigation/navigation_graph.xml
@@ -6,20 +6,6 @@
     app:startDestination="@id/toolBarFragment">
 
     <fragment
-        android:id="@+id/applicationCard"
-        android:name="com.example.marketplace_mvp.screens.feedscreen.ApplicationCard"
-        tools:layout="@layout/application_card"
-        android:label="ApplicationCard" />
-    <fragment
-        android:id="@+id/titleScreen"
-        android:name="com.example.marketplace_mvp.screens.TitleScreen"
-        tools:layout="@layout/title_screen"
-        android:label="TitleScreen" >
-        <action
-            android:id="@+id/action_titleScreen_to_toolBarFragment"
-            app:destination="@id/toolBarFragment" />
-    </fragment>
-    <fragment
         android:id="@+id/toolBarFragment"
         android:name="com.example.marketplace_mvp.screens.toolbar.ToolBarFragment"
         tools:layout="@layout/tool_bar_fragment"
@@ -28,4 +14,11 @@
             android:id="@+id/action_toolBarFragment_to_applicationCard"
             app:destination="@id/applicationCard" />
     </fragment>
+
+    <fragment
+        android:id="@+id/applicationCard"
+        android:name="com.example.marketplace_mvp.screens.feedscreen.ApplicationCard"
+        tools:layout="@layout/application_card"
+        android:label="ApplicationCard" />
+
 </navigation>
